Add buffer functionality to Gg class
The Gg class was updated to implement data buffering before sending it off to its endpoint. A buffer system was added to collect data and store it until the end of the script execution. This update aims to optimize data sending by reducing the number of requests sent out by using curl. It gathers the data and sends it off in chunks of 20. For every 20 pieces of data, it initiates a cURL request.

The `sendData()` method was replaced with `appendBuffer()` to collect data. This method was then called inside the `__destruct()` method, ensuring that data is sent at script termination. Every cURL request's timeout was also increased from 500ms to 1000ms.

The accompanying unit test was also updated to reflect this change. It now loops through a range of 1 to 100 to ensure the buffer has enough data to fill multiple chunks.