Add conditional check for exception listener activation
Added a new configuration option `gg.listeners.exception_handler_listener` which governs whether or not the ExceptionListener is active. This provides greater control to the user, allowing them turn off the exception listener selectively if required. A check is added in the handler to exit early if this config variable is not set.